+++ HISTORY.rst
@@ -4,6 +4,13 @@ History
 -------
 
 ------------------
+4.7.0 (2016-03-21)
+------------------
+
+ * Geant4 input to waveform simulator
+ * Tuning classification for XENON1T gas-mode zero-field.
+
+------------------
 4.6.1 (2016-03-07)
 ------------------